# CSE: inverse topological sort of letbindings and traverse once