@martijnbastiaan 2020-02-16
Not a review, but I'm not terribly excited about reintroducing fgl as a dependency. I removed it because it blocked a proper 8.8 release for Clash and it did so for several months (and possibly longer if we hadn't removed it), despite multiple attempts of contacting the project to release a new version. A quick look at the PR tracker for fgl seems to indicate that this has not changed.

I'm not going to try and veto it or anything, but it's something we need to consider.