@christiaanb 2020-02-16
If it happens again, we can either take over maintainership; or copy their code, it’s also bsd licensed.