@christiaanb 2020-02-16
I’ll try: http://hackage.haskell.org/package/containers-0.6.2.1/docs/Data-Graph.html 

See how it does