@christiaanb 2020-02-16
@martijnbastiaan Seems `Data.Graph` from `containers` does the job, so I removed the `fgl` dependency again.